requirements.txt
streamlit
pytest
google-generativeai
python-dotenv